backend/pyproject.toml
[project]
name = "musetub-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
	"alembic>=1.13",
	"asyncpg>=0.29",
	"fastapi>=0.110",
	"email-validator>=2.1",
	"python-jose[cryptography]>=3.3",
	"passlib[bcrypt]>=1.7",
	"bcrypt<5",
	"httpx>=0.27",
	"redis>=5.0",
	"sqlalchemy[asyncio]>=2.0",
	"uvicorn[standard]>=0.27",
	"pydantic>=2.6",
	"pydantic-settings>=2.2",
	"pytest>=8.0",
	"pytest-asyncio>=0.24",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
